--- a/packages/react-vapor/webpack.config.prod.minify.js
+++ b/packages/react-vapor/webpack.config.prod.minify.js
@@ -1,7 +1,0 @@
-const prodConfig = require('./webpack.config.prod');
-
-prodConfig.output.filename = '[name].min.js';
-prodConfig.mode = 'production';
-prodConfig.devtool = false;
-
-module.exports = prodConfig;