### Proposed Changes

I had a look at our build process and did a few optimizations:

- `build` command now produce the `.js`, `.min.js` and `.min.js.map` files in one single webpack execution
- `start` and `docs` command no longer generate the definition files (`transpileOnly: true`), but they still do the necessary type checking.
- better quality source maps are generated during the unit tests and while running `npm start`

#### Benchmarks

I timed each command once with linux `time` command on master and did the same process on this branch and here are the results. The time displayed here is the `real` time on my laptop so it takes into account the multi-threading, which means it could be slower on TravisCI or faster on your machine depending on your CPU specs.

| command | master branch | this branch |
| --- | --- | --- |
| start | 3 m 19 s | 30 s |
| docs | 3 m 7 s | 25 s |
| build | 56 s | 55 s |
| test | 2 m 30 s | 2 m 30 s |

### Potential Breaking Changes

- Source maps are now only generated for the minified files
- The files published to npm are now only the ones contained in the `dist` folder
- `minify` and `build:dev` commands have been removed

### Acceptance Criteria

-   [x] The proposed changes are covered by unit tests
-   [x] The potential breaking changes are clearly identified
-   [x] [README.md](https://github.com/coveo/react-vapor/blob/master/README.md) is adjusted to reflect the proposed changes (if relevant)